backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2

# LangGraph & LangChain
langgraph==0.2.45
langchain==0.3.7
langchain-core==0.3.15
langchain-groq==0.2.1
langsmith==0.1.143

# Vector Database
qdrant-client==1.12.0
sentence-transformers==3.2.1

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.12
httpx==0.27.2
aiofiles==24.1.0
requests==2.31.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
